relocated helper function file
