+++ client/src/components/rating_review/helperFunctions/helper.js
